Reduce zxing dependency in order to fix build issues
The 1.9.13 versions no longer seems to exist in a remote repository, at least any that I, or gradle, can easily find. They were stored in jCenter, which is deprecated, but now appear to be gone.
https://mvnrepository.com/artifact/me.dm7.barcodescanner/zxing

Looking at the github, it seems there aren't any significant features added after 1.9.8 that we need to use.
https://github.com/dm77/barcodescanner

I'm trying to build the Android version of Kalium and encountering build issues.